+++ tests/testapp/settings.py
@@ -30,6 +30,7 @@ MEDIA_ROOT = os.path.join(BASEDIR, 'media/')
 STATIC_ROOT = os.path.join(BASEDIR, 'static/')
 SECRET_KEY = 'supersikret'
 LOGIN_REDIRECT_URL = '/?login=1'
+ALLOWED_HOSTS = ['*']
 
 ROOT_URLCONF = 'testapp.urls'
 LANGUAGES = (('en', 'English'), ('de', 'German'))